Re-export more necessary types
